+++ include/types.h
@@ -1487,8 +1487,8 @@ typedef struct hc_device_param
   double  outerloop_msec;
   double  outerloop_multi;
 
-  u32     innerloop_pos;
-  u32     innerloop_left;
+  u64     innerloop_pos;
+  u64     innerloop_left;
 
   u32     exec_pos;
   double  exec_msec[EXEC_CACHE];
@@ -2779,8 +2779,8 @@ typedef struct device_info
   int     kernel_threads_dev;
   int     vector_width_dev;
   int     salt_pos_dev;
-  int     innerloop_pos_dev;
-  int     innerloop_left_dev;
+  u64     innerloop_pos_dev;
+  u64     innerloop_left_dev;
   int     iteration_pos_dev;
   int     iteration_left_dev;
   char   *device_name;
